chore: reorganize commentOperations.ts and update import paths
