+++ src/cli/update-strict-comments/__tests__/updateStrictComments.spec.ts
@@ -2,7 +2,7 @@ import { mocked } from 'jest-mock';
 import { isIgnoreCommentPresent, isStrictCommentPresent } from '../../isCommentPresent';
 import { getFilePathsWithErrors, getFilePathsOnPathWithoutErrors } from '../getFilePaths';
 import { updateStrictComments } from '../updateStrictComments';
-import { insertIgnoreComment, removeStrictComment } from '../commentOperations';
+import { insertIgnoreComment, removeStrictComment } from '../../commentOperations';
 
 jest.mock('../../findStrictErrors', () => ({
   findStrictErrors: jest.fn(),
